agrego swal, cambio estilos, optimizo codigo, archivo de example

+++ src/components/ItemCart.jsx
@@ -1,14 +1,16 @@
 import { Row, Col, Button, Image } from "react-bootstrap";
 import ItemTallesList from "./ItemTallesList";
 import { CartContext } from "../context/CartContext";
-import { useContext } from "react";
-import { useState } from "react";
+import { useContext, useState } from 'react'
+import Swal from 'sweetalert2'
 
 const ItemCart = ({ data }) => {
     const [boxShadow, setBoxShadow] = useState('');
-    const cantidad = data.selectStock.reduce((acumulador, elem) => acumulador + elem.quantity, 0);
-    const precioInicial = data.precio * cantidad;
-    const precioFinal = data.oferta ? precioInicial * data.porcDesc : precioInicial;
+    const unitInitialPrice = data.precio;
+    const unitFinalPrice = data.oferta ? unitInitialPrice * data.porcDesc : unitInitialPrice;
+    const quantity = data.selectStock.reduce((acumulador, elem) => acumulador + elem.quantity, 0);
+    const initialPrice = data.precio * quantity;
+    const finalPrice = data.oferta ? initialPrice * data.porcDesc : initialPrice;
     const { deleteTalleItemCart, deleteItemCart } = useContext(CartContext)
 
     const onDeleteTalle = (talleStock) => {
@@ -16,7 +18,15 @@ const ItemCart = ({ data }) => {
     }
 
     const onDelete = (id) => {
-        deleteItemCart(id);
+        Swal.fire({
+            title: "Seguro desea borrar la camiseta del carrito?",
+            showCancelButton: true,
+            confirmButtonText: "Eliminar",
+        }).then((result) => {
+            if (result.isConfirmed) {
+                deleteItemCart(id);
+            }
+        });
     }
 
     const handleHover = () => {
@@ -28,7 +38,7 @@ const ItemCart = ({ data }) => {
     };
 
     return (
-        <Row className="align-items-center mb-3 mx-0 px-3 border rounded p-2 itemCart"
+        <Row className="d-flex align-items-start mb-3 mx-0 px-3 border rounded p-2 itemCart"
             style={{
                 boxShadow: boxShadow,
                 transition: 'box-shadow 0.3s ease, transform 0.3s ease',
@@ -36,28 +46,34 @@ const ItemCart = ({ data }) => {
             onMouseOver={handleHover}
             onMouseLeave={handleLeave}
         >
-            <Col xs={4} md={4}>
+            <Col xs={5} md={4}>
                 <Image src={data.img} alt={data.nombre} fluid rounded />
             </Col>
-            <Col xs={5} md={5} className="d-flex flex-column justify-content-between">
+            <Col xs={4} md={5} className="d-flex flex-column justify-content-between">
                 <h5>{data.nombre}</h5>
-                <span className="mb-2 small text-muted">${data.precio}</span>
-
-                <ItemTallesList data={data.selectStock} handleDelete={onDeleteTalle}></ItemTallesList>
+                <ItemTallesList data={data.selectStock} handleDelete={onDeleteTalle} typeFlex={"flex-column"}></ItemTallesList>
             </Col>
             <Col xs={3} md={3} className="d-flex justify-content-center align-items-center flex-column ">
-                {data.oferta ? (
-                    <>
-                        <h4 className="text-decoration-line-through m-0">
-                            ${precioInicial}
-                        </h4>
-                        <h3 className="text-danger m-0">${precioFinal}</h3>
-                    </>
-                ) : (
-                    <h4>${precioFinal}</h4>
-                )}
 
-                <Button className="mt-4 btn-delete" size="xm"  onClick={() => onDelete(data.id)}>
+                {data.oferta && (
+                    <span className="text-decoration-line-through text-muted p-0 m-0 small">
+                        ${unitInitialPrice}
+                    </span>
+                )}
+                <span className="p-0 m-0 mb-1 text-muted small">
+                    ${unitFinalPrice}
+                </span>
+
+                {data.oferta && (
+                    <h5 className="text-decoration-line-through m-0">
+                        ${initialPrice}
+                    </h5>
+                )}
+                <h5 className={data.oferta ? "text-danger p-0 m-0" : "p-0 m-0"}>
+                    ${finalPrice}
+                </h5>
+
+                <Button className="mt-2 btn-delete" size="xm" onClick={() => onDelete(data.id)}>
                     Eliminar
                 </Button>
             </Col>
