+++ src/components/ItemDetail.jsx
@@ -1,15 +1,15 @@
 import { Container, Card, Row, Col, Button } from 'react-bootstrap';
 import ItemCount from './ItemCount';
-import { useContext } from 'react';
+import { useContext, useState } from 'react';
 import { CartContext } from '../context/CartContext';
 import TallesList from './TallesList';
-import { useState } from 'react';
 import ItemTallesList from './ItemTallesList';
 import { NavLink } from 'react-router-dom';
+import Swal from 'sweetalert2'
 
 const ItemDetail = ({ dataDetail }) => {
-    const precioInicial = dataDetail.precio;
-    const precioFinal = dataDetail.oferta ? precioInicial * dataDetail.porcDesc : precioInicial;
+    const initialPrice = dataDetail.precio;
+    const finalPrice = dataDetail.oferta ? initialPrice * dataDetail.porcDesc : initialPrice;
     const { addItemCart, deleteItemCart, itemCartStock } = useContext(CartContext);
     const [talle, setTalle] = useState('');
     const [stockDisponible, setStockDisponible] = useState('');
@@ -35,6 +35,12 @@ const ItemDetail = ({ dataDetail }) => {
     const onAddItem = () => {
         setItemAdd(true);
         addItemCart(dataDetail, arrayStock);
+
+        Swal.fire({
+            title: "Camiseta agregada!",
+            icon: "success",
+            draggable: true
+        });
     }
 
     const selectTalle = (btnValue) => {
@@ -45,10 +51,19 @@ const ItemDetail = ({ dataDetail }) => {
     }
 
     const onDelete = (id) => {
-        deleteItemCart(id);
-        setArrayStock([])
+
+        Swal.fire({
+            title: "Seguro desea borrar la camiseta del carrito?",
+            showCancelButton: true,
+            confirmButtonText: "Eliminar",
+        }).then((result) => {
+            if (result.isConfirmed) {
+                deleteItemCart(id);
+                setArrayStock([])
+            }
+        });
     }
-    
+
     const onDeleteTalle = (talleStock) => {
         const updateArrayStock = arrayStock.filter((elem) => elem.talle !== talleStock)
         setArrayStock(updateArrayStock)
@@ -71,17 +86,15 @@ const ItemDetail = ({ dataDetail }) => {
 
                             <Card.Text className="mb-2">
                                 Precio:&nbsp;
-                                {dataDetail.oferta ? (
-                                    <>
-                                        <span className="text-decoration-line-through">
-                                            ${precioInicial}
-                                        </span>
-                                        &nbsp;
-                                        <span className="text-danger">${precioFinal}</span>
-                                    </>
-                                ) : (
-                                    <span>${precioFinal}</span>
+
+                                {dataDetail.oferta && (
+                                    <span className="text-decoration-line-through p-0 m-0">
+                                        ${initialPrice}&nbsp;
+                                    </span>
                                 )}
+                                <span className={dataDetail.oferta ? "text-danger p-0 m-0" : "p-0 m-0"}>
+                                    ${finalPrice}
+                                </span>
                             </Card.Text>
 
                             <Card.Text className="mb-2">Temporada: {dataDetail.temporada}</Card.Text>
@@ -94,7 +107,7 @@ const ItemDetail = ({ dataDetail }) => {
                             {!itemAdd ?
                                 <>
                                     <ItemCount stock={stockDisponible} onAdd={onAddStock} talleSelect={talle} />
-                                    <ItemTallesList data={arrayStock} handleDelete={onDeleteTalle}></ItemTallesList>
+                                    <ItemTallesList data={arrayStock} handleDelete={onDeleteTalle} typeFlex={"flex-wrap"}></ItemTallesList>
                                     <Row className='d-flex justify-content-between mt-3'>
                                         <Col>
                                             <Button className='btn btn-dark' onClick={onAddItem} disabled={arrayStock.length === 0}>Confirmar</Button>
